Added P and I
The controller know have the ability to do proportional and integral
correction to the rotation, not to the speed. The speed is controlled
with a simple condition: if the distance is lower than a threshold stop,
if the distance is greater speed equals max velocity. It's a bit
simplistic, but it works so far.